move the test to integration
